server/requirements.txt
google-cloud-language
newsapi-python
flask
flask_cors
tensorflow
keras
nltk